pyproject.toml
[tool.poetry]
name = "fizzbuzzjazz-twitter-bot"
version = "0.1.0"
description = "Twitter bot for daily FizzBuzzJazz"
authors = ["Armand Rego <armand@weirdsheeplabs.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
